refactor(types): move shared types to shared/types folder
Create a new directory structure for shared types:
- Add bitrate.types.ts with CBR and VBR type definitions
- Add encoders.types.ts with AUDIO_ENCODER_NAMES, AUDIO_ENCODERS, and related Option types
- Add filters.types.ts with AUDIO_FILTER_NAMES, AUDIO_FILTERS, and Option types
- Add settings.types.ts with SettingsForm and Data types
- Add index.ts as barrel export for shared types
- Remove the now‑redundant type definitions from renderer's settings.types.ts
- Remove the Data type from root types.ts (now imported from @types)
