+++ src/shared/types/bitrate.types.ts
@@ -1,0 +1,19 @@
+export type CBR =
+	| "320k"
+	| "256k"
+	| "224k"
+	| "192k"
+	| "160k"
+	| "128k"
+	| "112k"
+	| "96k"
+	| "80k"
+	| "64k"
+	| "48k"
+	| "40k"
+	| "32k"
+	| "24k"
+	| "16k"
+	| "8k";
+
+export type VBR = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9";
